Hook up search API to use query string, part 1
