+++ src/components/blocks/SearchFilters.tsx
@@ -55,85 +55,87 @@ const SearchFilters: React.FC<SearchFiltersProps> = ({
     }
   }, [thisYear, searchCriteria]);
 
-  return <>
-    <div className="flex md:justify-between items-center mt-2 md:mt-0">
-      <div className="text-indigo-400 font-medium mr-2 md:mr-0">{t("Filter by")}</div>
-      {showClear && (
-        <button className="underline text-sm text-blue-500 hover:text-indigo-600 transition duration-300" onClick={handleClearSearch}>
-          Clear all filters
-        </button>
-      )}
-    </div>
+  return (
+    <>
+      <div className="flex md:justify-between items-center mt-2 md:mt-0">
+        <div className="text-indigo-400 font-medium mr-2 md:mr-0">{t("Filter by")}</div>
+        {showClear && (
+          <button className="underline text-sm text-blue-500 hover:text-indigo-600 transition duration-300" onClick={handleClearSearch}>
+            Clear all filters
+          </button>
+        )}
+      </div>
 
-    <div className="my-4 text-sm text-indigo-500">
-      <div>
-        <ExactMatch checked={searchCriteria.exact_match} id="exact-match" handleSearchChange={handleSearchChange} />
-      </div>
-      <div className="relative mt-6">
-        <BySelect
-          list={regions}
-          defaultValue={searchCriteria.keyword_filters?.regions ? searchCriteria.keyword_filters.regions[0] : ""}
-          onChange={handleRegionChange}
-          title={t("By region")}
-          keyField="slug"
-          keyFieldDisplay="display_value"
-          filterType="regions"
-        />
-      </div>
-      <div className="relative mt-6">
-        <ByTextInput
-          title={t("By country")}
-          list={filteredCountries}
-          selectedList={countryFilters}
-          keyField="slug"
-          keyFieldDisplay="display_value"
-          filterType="countries"
-          handleFilterChange={handleFilterChange}
-        />
-        <MultiList list={countryFilters} removeFilter={handleFilterChange} type="countries" />
-      </div>
-      <div className="relative mt-6 hidden">
-        <BySelect
-          list={sectors.filter((sector) => !sectorFilters.includes(sector.name))}
-          onChange={handleFilterChange}
-          title={t("By sector")}
-          keyField="name"
-          filterType="sectors"
-          defaultValue=""
-          defaultText={sectorFilters.length ? "Add more sectors" : "All"}
-        />
-        <MultiList list={sectorFilters} removeFilter={handleFilterChange} type="sectors" />
-      </div>
-      <div className="relative mt-8 mb-12">
-        <div className="">
-          <ByDateRange
-            title={t("By date range")}
-            type="year_range"
-            handleChange={handleYearChange}
-            defaultValues={searchCriteria.year_range}
-            min={minYear}
-            max={thisYear}
-            clear={showClear}
+      <div className="my-4 text-sm text-indigo-500">
+        <div>
+          <ExactMatch checked={searchCriteria.exact_match} id="exact-match" handleSearchChange={handleSearchChange} />
+        </div>
+        <div className="relative mt-6">
+          <BySelect
+            list={regions}
+            defaultValue={searchCriteria.keyword_filters?.regions ? searchCriteria.keyword_filters.regions[0] : ""}
+            onChange={handleRegionChange}
+            title={t("By region")}
+            keyField="slug"
+            keyFieldDisplay="display_value"
+            filterType="regions"
           />
         </div>
+        <div className="relative mt-6">
+          <ByTextInput
+            title={t("By country")}
+            list={filteredCountries}
+            selectedList={countryFilters}
+            keyField="slug"
+            keyFieldDisplay="display_value"
+            filterType="countries"
+            handleFilterChange={handleFilterChange}
+          />
+          <MultiList list={countryFilters} removeFilter={handleFilterChange} type="countries" />
+        </div>
+        <div className="relative mt-6 hidden">
+          <BySelect
+            list={sectors.filter((sector) => !sectorFilters.includes(sector.name))}
+            onChange={handleFilterChange}
+            title={t("By sector")}
+            keyField="name"
+            filterType="sectors"
+            defaultValue=""
+            defaultText={sectorFilters.length ? "Add more sectors" : "All"}
+          />
+          <MultiList list={sectorFilters} removeFilter={handleFilterChange} type="sectors" />
+        </div>
+        <div className="relative mt-8 mb-12">
+          <div className="">
+            <ByDateRange
+              title={t("By date range")}
+              type="year_range"
+              handleChange={handleYearChange}
+              defaultValues={searchCriteria.year_range}
+              min={minYear}
+              max={thisYear}
+              clear={showClear}
+            />
+          </div>
+        </div>
+        <div className="my-8 pt-8 border-t border-lineBorder">
+          <p className="text-center">
+            For more info see
+            <br />
+            {theme === "cpr" ? (
+              <ExternalLink url="https://github.com/climatepolicyradar/methodology" className="underline text-blue-600">
+                our methodology page
+              </ExternalLink>
+            ) : (
+              <Link href="/methodology" className="underline text-blue-600">
+                our methodology page
+              </Link>
+            )}
+          </p>
+        </div>
       </div>
-      <div className="my-8 pt-8 border-t border-lineBorder">
-        <p className="text-center">
-          For more info see
-          <br />
-          {theme === "cpr" ? (
-            <ExternalLink url="https://github.com/climatepolicyradar/methodology" className="underline text-blue-600">
-              our methodology page
-            </ExternalLink>
-          ) : (
-            <Link href="/methodology" className="underline text-blue-600">
-              our methodology page
-            </Link>
-          )}
-        </p>
-      </div>
-    </div>
-  </>;
+    </>
+  );
 };
 
 export default SearchFilters;
